Store loanIds in contract instead of loans (#14)

+++ src/components/BorrowPage.tsx
@@ -78,7 +78,7 @@ const BorrowPage = (props: any) => {
                         type="number"
                         label="Loan Amount (ETH)"
                         value={loanAmount}
-                        onChange={e => setLoanAmount(parseInt(e.target.value))}
+                        onChange={e => setLoanAmount(parseFloat(e.target.value))}
                     />
                     <TextField
                         fullWidth
